requirements.txt
Babel==2.11.0
beautifulsoup4==4.12.3
langchain==0.3.14
openai==1.59.7
pandas==1.5.3
protobuf==3.20.3
pymupdf4llm==0.0.17
python-dotenv==1.0.1
Requests==2.32.3
selenium==4.27.1
webdriver_manager==4.0.2
